patch aus confix-2.3.8.ebuild: confix-2.3.5-default-no-undefined.patch
Windows does not allow undefined symbols in shared libraries,
and this is bad practice anyway.  In fact, I've yet to see some
library that does require undefined symbols, which has not been
ported to Windows in one or the other way.
Enabling -no-undefined by default is the logical next step.

Change-Id: Idc870e3dcc2b8c1da0c97d0df62e4a5142339318